+++ docs/changelog.md
@@ -6,6 +6,12 @@ nav_order: 2
 
 # Changelog
 
+## Unreleased changes
+
+### Added
+
+- Convenience functions for KNX group communication (`xknx.tools`)
+
 ## 1.0.2 Route-back reconnect 2022-08-31
 
 ### Bugfixes